8511: Bring isReindexComplete into the lambda [skip ci]
