+++ webpack.config.lambda.migration.js
@@ -1,5 +1,4 @@
 const config = require('./webpack.config.lambda');
-const CopyPlugin = require('copy-webpack-plugin');
 
 module.exports = {
   ...config,
@@ -16,14 +15,4 @@ module.exports = {
     libraryTarget: 'umd',
     path: __dirname + '/web-api/migration-terraform/main/lambdas/dist',
   },
-  plugins: [
-    new CopyPlugin({
-      patterns: [
-        {
-          from: './shared/admin-tools/elasticsearch/check-reindex-complete.js',
-          to: '.',
-        },
-      ],
-    }),
-  ],
 };
